+++ CHANGELOG.md
@@ -3,10 +3,11 @@
 ## unreleased version -- v0.11.3
 ### Added
 - New endpoint to update whole collection JSON objects in the API response
+- Publication rights data object to collectionFetch API response
+### Fixed
 - Add new DB entities for automatic collections
 - Add support for 'Most Recent' automatic collection
 - Supported ES backed automatic collections
-### Fixed
 
 ## 2023-02-02 -- v0.11.2
 ### Added